refactored compatibility issues
