+++ app/components/ui/BurgerMenu/BurgerMenu.module.scss
@@ -30,8 +30,8 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  border: 1px solid var(--gray-secondary, #e2e6e9);
-  border-bottom: none;
+  border: none;
+  border: 1px solid var(--border-clr, #e2e6e9);
   width: 48px;
   height: 48px;
   background: none;
