+++ app/components/ui/HeaderActions.tsx/HeaderActions.module.scss
@@ -11,31 +11,18 @@
   justify-content: center;
   width: 48px;
   height: 100%;
-  border: 1px solid var(--gray-secondary, #e2e6e9);
+  border-left: 1px solid var(--border-clr, #e2e6e9);
   text-decoration: none;
-  background: none;
+
   cursor: pointer;
   transition: transform 0.2s ease;
-
-  &:hover {
-    transform: scale(1.1);
-  }
 
   @media screen and (min-width: 640px) {
     display: flex;
   }
-}
 
-.icon_border {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 64px;
-  height: 64px;
-  border: 1px solid var(--gray-secondary, #e2e6e9);
-
-  &:hover {
-    transform: scale(1.1);
+  &:last-child {
+    border-right: none;
   }
 }
 
@@ -43,8 +30,8 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  border: 1px solid var(--gray-secondary, #e2e6e9);
-  border-bottom: none;
+  border: none;
+  border-left: 1px solid var(--border-clr, #e2e6e9);
   width: 48px;
   height: 48px;
   background: none;
